Bug 1983320 ML-DSA tests for Sign/Verify, certificates and pkcs12
These tests use certutil and pk12util to verify that we can generate and verify certificates using the cryptohi signing/verify interfaces. You'll need the main set of ml-dsa patches as well as a freebl interface for these to run, so the can't be checked in until all the rest has landed.

Differential Revision: https://phabricator.services.mozilla.com/D264713